fix(deps): 收敛公共库升级与执行边界
升级 Go 与公共库依赖并同步 CI、Docker 和当前版本文档。

适配 toolkit v0.3 执行契约，以原子候选发布沙箱策略，并在缺少目的地址过滤时拒绝宿主网络。

同步 SSE、SSRF、重试和文件身份边界的回归测试。

+++ skill/builtin/code_exec_file_identity_darwin.go
@@ -13,9 +13,10 @@ func codeExecPlatformFileIdentity(file *os.File, info os.FileInfo) (codeExecPlat
 		return codeExecPlatformIdentity{}, errCodeExecFileIdentityUnavailable
 	}
 	return codeExecPlatformIdentity{
-		Volume:           uint64(stat.Dev),
+		// Darwin 的 dev_t 是有符号 32 位设备号；这里有意保留内核位模式后再拓宽。
+		Volume:           uint64(uint32(stat.Dev)), // #nosec G115 -- 按设备号位模式转换，不是数值符号转换。
 		FileIDHigh:       0,
-		FileIDLow:        uint64(stat.Ino),
+		FileIDLow:        stat.Ino,
 		Links:            uint64(stat.Nlink),
 		ChangeTimeSec:    stat.Ctimespec.Sec,
 		ChangeTimeNsec:   stat.Ctimespec.Nsec,
